+++ frontend/src/app/[lang]/blog/[category]/[slug]/layout.tsx
@@ -1,103 +1,104 @@
-import React from "react";
+import React from 'react';
 
-import { fetchAPI } from "@/app/utils/fetch-api";
-import ArticleSelect from "@/app/components/ArticleSelect";
-
+import ArticleSelect from '@/app/[lang]/components/ArticleSelect';
+import { fetchAPI } from '@/app/[lang]/utils/fetch-api';
 
 async function fetchSideMenuData(filter: string) {
-  try {
-    const token = process.env.NEXT_PUBLIC_STRAPI_API_TOKEN;
-    const options = { headers: { Authorization: `Bearer ${token}` } };
+    try {
+        const token = process.env.NEXT_PUBLIC_STRAPI_API_TOKEN;
+        const options = { headers: { Authorization: `Bearer ${token}` } };
 
-    const categoriesResponse = await fetchAPI("/categories", { populate: "*" }, options);
+        const categoriesResponse = await fetchAPI('/categories', { populate: '*' }, options);
 
-    const articlesResponse = await fetchAPI(
-      "/articles",
-      filter
-        ? {
-            filters: {
-              category: {
-                name: filter,
-              },
-            },
-          }
-        : {},
-      options
-    );
+        const articlesResponse = await fetchAPI(
+            '/articles',
+            filter
+                ? {
+                      filters: {
+                          category: {
+                              name: filter,
+                          },
+                      },
+                  }
+                : {},
+            options
+        );
 
-    return {
-      articles: articlesResponse.data,
-      categories: categoriesResponse.data,
-    };
-  } catch (error) {
-    console.error(error);
-  }
+        return {
+            articles: articlesResponse.data,
+            categories: categoriesResponse.data,
+        };
+    } catch (error) {
+        console.error(error);
+    }
 }
 
 interface Category {
-  id: number;
-  attributes: {
-    name: string;
-    slug: string;
-  };
+    id: number;
+    attributes: {
+        name: string;
+        slug: string;
+    };
 }
 
 interface Article {
-  id: number;
-  attributes: {
-    title: string;
-    slug: string;
-  };
+    id: number;
+    attributes: {
+        title: string;
+        slug: string;
+    };
 }
 
 interface Data {
-  articles: Article[];
-  categories: Category[];
+    articles: Article[];
+    categories: Category[];
 }
 
 export default async function LayoutRoute({
-  params,
-  children,
+    params,
+    children,
 }: {
-  children: React.ReactNode;
-  params: {
-    slug: string;
-    category: string;
-  };
+    children: React.ReactNode;
+    params: {
+        slug: string;
+        category: string;
+    };
 }) {
+    const { category } = params;
+    const { categories, articles } = (await fetchSideMenuData(category)) as Data;
 
-  const { category } = params;
-  const { categories, articles } = (await fetchSideMenuData(category)) as Data;
-
-  return (
-    <section className="container p-8 mx-auto space-y-6 sm:space-y-12">
-      <div className="grid grid-cols-1 md:grid-cols-3 gap-2 lg:gap-4">
-        <div className="col-span-2">{children}</div>
-        <aside>
-          <ArticleSelect categories={categories} articles={articles} params={params}/>
-        </aside>
-      </div>
-    </section>
-  );
+    return (
+        <section className="container p-8 mx-auto space-y-6 sm:space-y-12">
+            <div className="grid grid-cols-1 md:grid-cols-3 gap-2 lg:gap-4">
+                <div className="col-span-2">{children}</div>
+                <aside>
+                    <ArticleSelect categories={categories} articles={articles} params={params} />
+                </aside>
+            </div>
+        </section>
+    );
 }
 
-
 export async function generateStaticParams() {
-  const token = process.env.NEXT_PUBLIC_STRAPI_API_TOKEN;
-  const path = `/articles`;
-  const options = { headers: { Authorization: `Bearer ${token}` } };
-  const articleResponse = await fetchAPI(path, {
-    populate: ['category']
-  }, options);
+    const token = process.env.NEXT_PUBLIC_STRAPI_API_TOKEN;
+    const path = `/articles`;
+    const options = { headers: { Authorization: `Bearer ${token}` } };
+    const articleResponse = await fetchAPI(
+        path,
+        {
+            populate: ['category'],
+        },
+        options
+    );
 
-  return articleResponse.data.map(
-    (article: {
-      attributes: {
-        slug: string;
-        category: {
-          slug: string,
-        }
-      };
-    }) => ({ slug: article.attributes.slug, category: article.attributes.slug })
-  );
+    return articleResponse.data.map(
+        (article: {
+            attributes: {
+                slug: string;
+                category: {
+                    slug: string;
+                };
+            };
+        }) => ({ slug: article.attributes.slug, category: article.attributes.slug })
+    );
 }
